toolkit/components/passwordmgr/test/mochitest/test_autocomplete_new_password.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofill and autocomplete on autocomplete=new-password fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="pwmgr_common.js"></script>
  <script src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: autofill with autocomplete=new-password fields

<script>
let chromeScript = runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  login1.init("https://example.com", "https://autofill", null,
              "user1", "pass1", "", "");

  Services.logins.addLogin(login1);
});

let readyPromise = registerRunTests();
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 is the reference, sanity-check -->
  <form id="form1" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="p">
    <button type="submit">Submit</button>
  </form>

  <!-- form2 uses a new-password type=password field -->
  <form id="form2" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
const {ContentTaskUtils} =
  SpecialPowers.Cu.import("resource://testing-common/ContentTaskUtils.jsm", {});

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;
});

add_task(async function test_autofillAutocompleteUsername_noGeneration() {
  // reference form was filled as expected?
  checkForm(1, "user1", "pass1");

  // 2nd form should not be filled
  checkForm(2, "", "");

  $_(2, "uname").focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkForm(2, "user1", "pass1");

  document.getElementById("form2").reset();
});

add_task(async function test_autofillAutocompletePassword_noGeneration() {
  // 2nd form should not be filled
  checkForm(2, "", "");

  let pword = $_(2, "pword");
  pword.focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  // Can't use promiseFormsProcessed() when autocomplete fills the field directly.
  await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
  checkForm(2, "", "pass1");

  // No autocomplete results should appear for non-empty pw fields.
  synthesizeKey("KEY_ArrowDown");
  await promiseNoUnexpectedPopupShown();

  document.getElementById("form2").reset();
});

// All tests below this are with generation prefs enabled.

add_task(async function test_autofillAutocompleteUsername_noGeneration() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.generation.available", true],
    ["signon.generation.enabled", true],
    ["signon.showAutoCompleteOrigins", true],
  ]});

  // 2nd form should not be filled
  checkForm(2, "", "");

  $_(2, "uname").focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  // No generation option on username fields.
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkForm(2, "user1", "pass1");

  document.getElementById("form2").reset();
});

add_task(async function test_autofillAutocompletePassword_withGeneration() {
  // 2nd form should not be filled
  checkForm(2, "", "");

  let pword = $_(2, "pword");
  pword.focus();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = [
    "user1",
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  // Can't use promiseFormsProcessed() when autocomplete fills the field directly.
  await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
  checkForm(2, "", "pass1");

  // No autocomplete results should appear for non-empty pw fields.
  synthesizeKey("KEY_ArrowDown");
  await promiseNoUnexpectedPopupShown();

  while (pword.value) {
    synthesizeKey("KEY_Backspace");
  }

  info("This time select the generated password");
  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  results = await shownPromise;
  expectedACLabels = [
    "user1",
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
  let generatedPW = pword.value;
  is(generatedPW.length, GENERATED_PASSWORD_LENGTH, "Check generated password length");
  ok(generatedPW.match(GENERATED_PASSWORD_REGEX), "Check generated password format");

  while (pword.value) {
    synthesizeKey("KEY_Backspace");
  }

  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  results = await shownPromise;
  expectedACLabels = [
    "user1",
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
  // Same generated password should be used.
  checkForm(2, "", generatedPW);

  document.getElementById("form2").reset();
});
</script>
</pre>
</body>
</html>
